% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_constructors.R
\name{CUSTOM_data}
\alias{CUSTOM_data}
\title{Creates a \code{CUSTOM_data} object}
\usage{
CUSTOM_data(dat, keep_rownames_as = NULL)
}
\arguments{
\item{dat}{Tibble of custom genomic features. See Details.}

\item{keep_rownames_as}{Character, the name of the column in which to save the
rownames of the input data-frame. Default value is \code{NULL}, i.e. rownames will
be discarded.}
}
\value{
A \code{CUSTOM_data} object, i.e. a tibble.
}
\description{
Creates a \code{CUSTOM_data} object from a tibble or data-frame of custom genomic features.
}
\details{
The input data should have one row per marker, and at least the
following columns:
\itemize{
\item \code{chromosome}: character column, chromosome on which the feature is located.
\item \code{position}: numeric, the physical position of the feature along the chromosome (in bp) OR
\item \code{start} and \code{end}: numeric, starting and end position of the feature
(in bp). If there is no \code{position} column, it will be constructed as the
middle value (mean) between \code{start} and \code{end}. If there is a \code{position}
column but no \code{start} and/or \code{end} columns, these will be set to the values
in the \code{position} column.
\item \code{score}: numeric, score to be used for the genomic feature.
}
}
\examples{
x <- get_example_data()

CUSTOM_data(x[["GWAS"]])
}
