% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_constructors.R
\name{DE_data}
\alias{DE_data}
\title{Creates a DE_data object}
\usage{
DE_data(dat, keep_rownames_as = NULL)
}
\arguments{
\item{dat}{Tibble, results from a GWAS analysis. See Details.}

\item{keep_rownames_as}{Character, the name of the column in which to save the
rownames of the input data-frame. Default value is \code{NULL}, i.e. rownames will
be discarded.}
}
\value{
A \code{DE_data} object, i.e. a tibble.
}
\description{
The input data should have one row per gene or transcript, and at least the
following columns:
}
\details{
\itemize{
\item \code{chromosome}: character column, chromosome on which the gene/transcript is located.
\item \code{score}: numeric, the DE score of the gene/transcript.
\item \code{start}: numeric, starting position of the gene/transcript (in bp).
\item \code{end}: numeric, end position of the gene/transcript (in bp).
}
}
