% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hidecan_plot.R
\name{.add_data_type}
\alias{.add_data_type}
\title{Add data type to HIDECAN plot}
\usage{
.add_data_type(
  p,
  toplot,
  i,
  paes,
  add_new_legend,
  point_size,
  label_size,
  label_padding
)
}
\arguments{
\item{p}{Current ggplot.}

\item{toplot}{Tibble of data to plot.}

\item{i}{Character, data type to add in the plot.}

\item{paes}{Named list, aesthetics to use for this data type (see
\code{\link[=hidecan_aes]{hidecan_aes()}} for information about the required values).}

\item{add_new_legend}{Logical, whether the current data type requires a new
legend for fill.}

\item{point_size}{Numeric, size of the points in the plot.}

\item{label_size}{Numeric, size of the gene labels in the plot.}

\item{label_padding}{Numeric, amount of padding around gene labels in the
plot, as unit or number.}
}
\value{
Ggplot \code{p} to which the new data type has been added.
}
\description{
Add data type to HIDECAN plot
}
