% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hidecan_plot.R
\name{.check_chrom_limits}
\alias{.check_chrom_limits}
\title{Check chromosome limits}
\usage{
.check_chrom_limits(chrom_limits, chrom_length, chroms)
}
\arguments{
\item{chrom_limits}{Integer vector of length 2, or named list where the
elements are integer vectors of length 2. If vector, gives the lower and
upper limit of the chromosomes (in bp) to use in the plot. If a named list,
names should correspond to chromosome names. Gives for each chromosome the
lower and upper limits (in bp) to use in the plot. Doesn't have to be
specified for all chromosomes. Default value is \code{NULL}, i.e. no limits are
applied to the chromosomes (they will be plotted in their entirety).}

\item{chrom_length}{Tibble of chromosome length.}

\item{chroms}{Character vector, name of chromosomes to include in the plot.}
}
\value{
Tibble of chromosome limits to use in the plot.
}
\description{
Check chromosome limits
}
