% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hidecan_plot.R
\name{.check_chroms}
\alias{.check_chroms}
\title{Check chromosomes to plot}
\usage{
.check_chroms(toplot, chrom_length, chroms, remove_empty_chrom)
}
\arguments{
\item{toplot}{Tibble, main data-frame for plot.}

\item{chrom_length}{Tibble of chromosome length.}

\item{chroms}{Character vector, name of chromosomes to include in the plot.
If \code{NULL} (default value), all chromosomes will be included.}

\item{remove_empty_chrom}{Logical, should chromosomes with no significant
markers/genes nor candidate genes be removed from the plot? Default value
if \code{FALSE}.}
}
\value{
Character vector of chromosomes to plot.
}
\description{
Check chromosomes to plot
}
