% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_cols}
\alias{.check_cols}
\title{Checks whether some columns are present in a tibble}
\usage{
.check_cols(x, col_names, param_name = "Input data-frame")
}
\arguments{
\item{x}{Tibble}

\item{col_names}{character vector of column names}

\item{param_name}{Character, name of the dataframe to use in the error message.}
}
\value{
invisible NULL
}
\description{
Checks whether some columns are present in a tibble
}
