% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_chromosomes_length.R
\name{.compute_chrom_length_genes}
\alias{.compute_chrom_length_genes}
\title{Computes chromosomes' length for a tibble of genes}
\usage{
.compute_chrom_length_genes(x)
}
\arguments{
\item{x}{Either a \code{DE_data} or \code{CAN_data} object.}
}
\value{
A tibble with two columns: \code{chrom} (chromosome name) and
\code{length} (chromosome length in base pair).
}
\description{
Computes the length (in bp) of each chromosome as the maximum
position of genes on the chromosome.
}
