% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_chromosomes_length.R
\name{.compute_chrom_length_markers}
\alias{.compute_chrom_length_markers}
\title{Computes chromosomes' length for a tibble of markers}
\usage{
.compute_chrom_length_markers(x)
}
\arguments{
\item{x}{Either a \code{GWAS_data} or \code{CUSTOM_data} object.}
}
\value{
A tibble with two columns: \code{chromosome} (chromosome name) and
\code{length} (chromosome length in base pair).
}
\description{
Computes the length (in bp) of each chromosome as the maximum
position of chromosomes on the chromosome.
}
