% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hidecan_plot.R
\name{.get_aes_type}
\alias{.get_aes_type}
\title{Returns either "aes_type" attribute or object class}
\usage{
.get_aes_type(x)
}
\arguments{
\item{x}{A \code{GWAS_data_thr}, \code{DE_data_thr}, \code{CAN_data_thr} or
\code{CUSTOM_data_thr} object.}
}
\value{
Either "aes_type" attribute or object class.
}
\description{
Returns either "aes_type" attribute or object class
}
