% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hidecan_plot.R
\name{.get_plot_aes}
\alias{.get_plot_aes}
\title{Computes plot aesthetics}
\usage{
.get_plot_aes(aes_types, colour_genes_by_score, custom_aes)
}
\arguments{
\item{aes_types}{Character vector of data types (one per dataset to plot),
should be computed with \code{\link[=.get_aes_type]{.get_aes_type()}}.}

\item{colour_genes_by_score}{Logical, whether to colour the genes by score
(\code{TRUE}) or by log2(fold-change) (\code{FALSE}).}

\item{custom_aes}{Named list of plot aesthetics for custom data types. See
\code{\link[=hidecan_aes]{hidecan_aes()}} for information about the content of each element.}
}
\value{
Named list of plot aesthetics.
}
\description{
Get the list of aesthetics for each data types in the plot.
}
