% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gwaspoly_example_data}
\alias{get_gwaspoly_example_data}
\title{Get set of GWASpoly example data}
\usage{
get_gwaspoly_example_data(with_thresholds = TRUE)
}
\arguments{
\item{with_thresholds}{Logical, should the GWASpoly object returned
contain significance threshold? Default value is \code{TRUE}.}
}
\value{
\itemize{
\item if \code{with_thresholds} is \code{TRUE}: a \code{GWASpoly.thresh} object
(returned by the \code{GWASpoly::set.threshold()} function).
if \code{with_thresholds} is \code{FALSE}: a \code{GWASpoly.fitted} object
(returned by the \code{GWASpoly::GWASpoly()} function).
}
}
\description{
Returns a list of GWASpoly example datasets.
}
