% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hidecan_plot.R
\name{hidecan_plot}
\alias{hidecan_plot}
\title{Wrapper to create a HIDECAN plot}
\usage{
hidecan_plot(
  gwas_list = NULL,
  de_list = NULL,
  can_list = NULL,
  score_thr = c(GWAS = 4, DE = 1.3),
  log2fc_thr = 1,
  chrom_length = NULL,
  colour_genes_by_score = TRUE,
  title = NULL,
  subtitle = NULL,
  n_rows = NULL,
  n_cols = 2,
  legend_position = "bottom",
  point_size = 3,
  label_size = 3.5,
  label_padding = 0.15
)
}
\arguments{
\item{gwas_list}{Data-frame or list of data-frames containing GWAS results,
each with at least a \code{chromosome}, \code{position} and either \code{padj} or \code{score}
columns. If a named list, the names will be used in the plot.}

\item{de_list}{Data-frame or list of data-frames containing DE results,
each with at least a \code{chromosome}, \code{start}, \code{end}, \code{log2FoldChange} and
either \code{padj} or \code{score} columns. If a named list, the names will be used in
the plot.}

\item{can_list}{Data-frame or list of data-frames containing candidate genes,
each with at least a \code{chromosome}, \code{start}, \code{end} and \code{name} columns. If a
named list, the names will be used in the plot.}

\item{score_thr}{Named numeric vector of length 2, with names \code{'GWAS'} and
\code{'DE'}. The score threshold for GWAS and DE results that will be used to
select which markers or genes will be plotted.}

\item{log2fc_thr}{Numeric, the log2(fold-change) threshold that will be used
to select which genes will be plotted.}

\item{chrom_length}{Optional, tibble with columns \code{chromosome} and \code{length},
giving for each chromosome its length in bp. If \code{NULL} (the default), will
be inferred from the GWAS, DE and candidate gene data.}

\item{colour_genes_by_score}{Logical, whether to colour the genes by score
(\code{TRUE}) or by log2(fold-change) (\code{FALSE}). Default value is \code{TRUE}.}

\item{title}{Character, title of the plot. Default value is \code{NULL} (i.e.
no title will be added to the plot).}

\item{subtitle}{Character, subtitle of the plot. Default value is \code{NULL}
(i.e. no subtitle will be added to the plot).}

\item{n_rows}{Integer, number of rows of chromosomes to create in the plot.
Default value is \code{NULL}.}

\item{n_cols}{Integer, number of columns of chromosomes to create in the plot.
Default value is 2.}

\item{legend_position}{Character, position of the legend in the plot. Can be
\code{bottom} (default value), \code{top}, \code{right}, \code{left} or \code{none}.}

\item{point_size}{Numeric, size of the points in the plot. Default value is 3.}

\item{label_size}{Numeric, size of the gene labels in the plot. Default value is
3.5 (for \code{\link[ggrepel]{geom_label_repel}}).}

\item{label_padding}{Numeric, amount of padding around gene labels in the plot,
as unit or number. Default value is 0.15
(for \link[ggrepel]{geom_label_repel}).}
}
\value{
a ggplot.
}
\description{
Wrapper function to create a HIDECAN plot from GWAS results,
DE results or candidate genes.
}
