% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwaspoly_helper_functions.R
\name{hidecan_plot_from_gwaspoly}
\alias{hidecan_plot_from_gwaspoly}
\title{Creates a HIDECAN plot from GWASpoly output}
\usage{
hidecan_plot_from_gwaspoly(gwaspoly_output, traits = NULL, models = NULL, ...)
}
\arguments{
\item{gwaspoly_output}{A \code{GWASpoly.thresh} object
(returned by the \code{GWASpoly::set.threshold()} function).}

\item{traits}{Character vector, traits for which GWAS results should be
extracted. If \code{NULL} (default value), all traits present are considered.}

\item{models}{Character vector, genetic models for which GWAS results should be
extracted. If \code{NULL} (default value), all genetic models present are considered.}

\item{...}{Further arguments passed to the \code{\link{create_hidecan_plot}()}
function.}
}
\value{
A ggplot.
}
\description{
Creates a HIDECAN plot from the GWAS results from GWASpoly.
}
