% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Creates a Manhattan plot}
\usage{
manhattan_plot(
  gwas_df,
  score_thr = NULL,
  chroms = NULL,
  title = NULL,
  subtitle = NULL,
  size_range = c(1, 3),
  chrom_col = NULL
)
}
\arguments{
\item{gwas_df}{Data-frame of GWAS results, with at least a \code{chromosome},
\code{position} and either \code{padj} or \code{score} columns.}

\item{score_thr}{Numeric, the significance threshold on GWAS score. If a value
is provided, will be represented in the Manhattan plot as a red dashed line. If
\code{NULL} (default value), no significance threshold line will be drawn.}

\item{chroms}{Character vector, name of chromosomes to include in the plot.
If \code{NULL} (default value), all chromosomes will be included.}

\item{title}{Character, title of the plot. Default value is \code{NULL} (i.e.
no title will be added to the plot).}

\item{subtitle}{Character, subtitle of the plot. Default value is \code{NULL}
(i.e. no subtitle will be added to the plot).}

\item{size_range}{Numeric vector of length 2, the minimum and maximum point
size in the plot. Points size is proportional to their GWAS score. Default
value is \code{c(1, 3)}.}

\item{chrom_col}{Character vector of colour names or code, colours to use to draw
the points for each chromosome. Names will be ignored. If vector provided
contains less colours than the number of chromosomes to plot, the values will
be recycled. If \code{NULL}, default, colours will be automatically chosen from a
predefined palette.}
}
\value{
A \code{ggplot}.
}
\description{
Creates a Manhattan plot from a data-frame of GWAS results.
}
\examples{
if (interactive()){
x <- get_example_data()[["GWAS"]]

manhattan_plot(x)

## Adding a significance threshold line in the plot
manhattan_plot(x, score_thr = 4)


## Use only two colours for the chromosomes
manhattan_plot(x,
               score_thr = 4,
               chrom_col = c("dodgerblue1", "dodgerblue4"))
}
}
