% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_constructors.R
\name{new_DE_data}
\alias{new_DE_data}
\title{\code{DE_data} constructor}
\usage{
new_DE_data(dat)
}
\arguments{
\item{dat}{Tibble, results from a differential expression analysis, with at least columns
\code{chromosome}, \code{score}, \code{log2FoldChange}, \code{start}, \code{end} and \code{position}.}
}
\value{
A \code{DE_data} object, i.e. a tibble.
}
\description{
\code{DE_data} constructor
}
