% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_constructors.R
\name{new_GWAS_data}
\alias{new_GWAS_data}
\title{GWAS_data constructor}
\usage{
new_GWAS_data(dat, title = "GWAS")
}
\arguments{
\item{dat}{Tibble, results from a GWAS analysis, with at least columns
\code{chromosome}. \code{position} and \code{score}.}

\item{title}{Character, the name of the analysis. Default value is \code{"GWAS"}.}
}
\value{
A \code{GWAS_data} object, i.e. a tibble.
}
\description{
GWAS_data constructor
}
