% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_constructors.R
\name{new_QTL_data}
\alias{new_QTL_data}
\title{\code{QTL_data} constructor}
\usage{
new_QTL_data(dat)
}
\arguments{
\item{dat}{Tibble, results from a QTL mapping, with at least columns
\code{chromosome}, \code{start}, \code{end} and \code{score}.}
}
\value{
A \code{QTL_data} object, i.e. a tibble.
}
\description{
\code{QTL_data} constructor
}
