% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_constructors.R
\name{QTL_data}
\alias{QTL_data}
\title{Creates a \code{QTL_data} object}
\usage{
QTL_data(dat, keep_rownames_as = NULL)
}
\arguments{
\item{dat}{Tibble, results from a QTL mapping analysis. See Details.}

\item{keep_rownames_as}{Character, the name of the column in which to save
the rownames of the input data-frame. Default value is \code{NULL}, i.e.
rownames will be discarded.}
}
\value{
A \code{QTL_data} object, i.e. a tibble.
}
\description{
Creates a \code{QTL_data} object from a tibble or data-frame of QTL mapping
results.
}
\details{
The input data should have one row per QTL region, and at least the following
columns:
\itemize{
\item \code{chromosome}: character column, chromosome on which the QTL region is located.
\item \code{start} and \code{end}: numeric, starting and end position of the QTL region
(in bp). A column \code{position} will be constructed as the middle value (mean)
between \code{start} and \code{end}.
\item \code{score} or \code{padj}: numeric, the QTL score or adjusted p-value of the
QTL region. If column \code{score} column is missing, will be constructed as
\code{-log10(padj)}.
}
}
\examples{
x <- get_example_data()

QTL_data(x[["QTL"]])
}
