% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{run_hidecan_shiny}
\alias{run_hidecan_shiny}
\title{Launches the HIDECAN shiny app}
\usage{
run_hidecan_shiny()
}
\value{
No return value, called for side effects (launching the shiny app).
}
\description{
Starts the HIDECAN shiny app. The app reads in csv data to produce
a HIDECAN plot.
}
