% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_constructors.R
\name{validate_CUSTOM_data}
\alias{validate_CUSTOM_data}
\title{Checks validity of input for \code{CUSTOM_data} constructor}
\usage{
validate_CUSTOM_data(x)
}
\arguments{
\item{x, }{a \code{CUSTOM_data} object constructed via \link{new_CUSTOM_data}.}
}
\value{
A \code{CUSTOM_data} object, i.e. a tibble.
}
\description{
Checks validity of input for \code{CUSTOM_data} constructor
}
