% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_constructors.R
\name{validate_QTL_data}
\alias{validate_QTL_data}
\title{Checks validity of input for \code{QTL_data} constructor}
\usage{
validate_QTL_data(x)
}
\arguments{
\item{x, }{a \code{QTL_data} object constructed via \link{new_QTL_data}.}
}
\value{
A \code{QTL_data} object, i.e. a tibble.
}
\description{
Checks validity of input for \code{QTL_data} constructor
}
