% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_constructors.R
\name{GWAS_data}
\alias{GWAS_data}
\title{Creates a GWAS_data object}
\usage{
GWAS_data(dat, keep_rownames_as = NULL)
}
\arguments{
\item{dat}{Tibble, results from a GWAS analysis. See Details.}

\item{keep_rownames_as}{Character, the name of the column in which to save the
rownames of the input data-frame. Default value is \code{NULL}, i.e. rownames will
be discarded.}
}
\value{
A \code{GWAS_data} object, i.e. a tibble.
}
\description{
The input data should have one row per marker, and at least the
following columns:
}
\details{
\itemize{
\item \code{chromosome}: character column, chromosome on which the marker is located.
\item \code{position}: numeric, the physical position of the marker along the chromosome (in bp).
\item \code{score}: numeric, the GWAS score of the marker.
}
}
