% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwaspoly_helper_functions.R
\name{hidecan_plot_from_gwaspoly}
\alias{hidecan_plot_from_gwaspoly}
\title{Create a hidecan plot from a GWASpoly output}
\usage{
hidecan_plot_from_gwaspoly(gwaspoly_input, ...)
}
\arguments{
\item{gwaspoly_input}{A \code{GWASpoly.thresh} object
(returned by the \code{GWASpoly::set.threshold()} function).}

\item{...}{Further arguments passed to the \code{\link{create_hidecan_plot}()}
function.}
}
\value{
A ggplot.
}
\description{
Creates a hidecan plot from the GWAS results from GWASpoly.
}
\examples{
\dontrun{
x <- get_gwaspoly_example_data(with_thresholds = TRUE)

hidecan_plot_from_gwaspoly(x)
}
}
