% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_constructors.R
\name{new_CAN_data}
\alias{new_CAN_data}
\title{CAN_data constructor}
\usage{
new_CAN_data(dat)
}
\arguments{
\item{dat}{Tibble, containing information about genes of interest, with at least columns
\code{chromosome}, \code{start}, \code{end} and \code{name}.}
}
\value{
A \code{CAN_data} object, i.e. a tibble.
}
\description{
CAN_data constructor
}
